requirements.txt
openai
rouge-score
nltk
python-dotenv
ollama
scikit-learn
